refactor(issue-33): added resource specific command for content management
